update fiile structure
