+++ src/scenes/Activities/scenes/EventHistory.js
@@ -21,14 +21,14 @@ import cx from 'classnames';
 import { FormattedMessage, injectIntl, intlShape, defineMessages } from 'react-intl';
 
 import styles from './styles';
-import Config from '../../../../config/app';
-import TransactionHistoryID from '../../../../components/TransactionHistoryAddressAndID/id';
-import TransactionHistoryAddress from '../../../../components/TransactionHistoryAddressAndID/address';
-import graphqlActions from '../../../../redux/Graphql/actions';
-import { getShortLocalDateTimeString, getDetailPagePath } from '../../../../helpers/utility';
-import { i18nToUpperCase } from '../../../../helpers/i18nUtil';
-import { getTxTypeString } from '../../../../helpers/stringUtil';
-import { TransactionType, SortBy, AppLocation } from '../../../../constants';
+import Config from '../../../config/app';
+import TransactionHistoryID from '../../../components/TransactionHistoryAddressAndID/id';
+import TransactionHistoryAddress from '../../../components/TransactionHistoryAddressAndID/address';
+import graphqlActions from '../../../redux/Graphql/actions';
+import { getShortLocalDateTimeString, getDetailPagePath } from '../../../helpers/utility';
+import { i18nToUpperCase } from '../../../helpers/i18nUtil';
+import { getTxTypeString } from '../../../helpers/stringUtil';
+import { TransactionType, SortBy, AppLocation } from '../../../constants';
 
 
 const messages = defineMessages({ // eslint-disable-line
