--- src/scenes/Activities/scenes/Withdraw/styles.js
@@ -1,16 +1,0 @@
-const styles = () => ({
-  row: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    width: '100%',
-    margin: 20,
-  },
-  loading: {
-    width: '5rem',
-    height: '5rem',
-    marginTop: 360,
-  },
-});
-
-export default styles;
