+++ src/scenes/Activities/scenes/styles.js
